requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==5.0.3
django-cache-url==3.4.5
environs==11.0.0
marshmallow==3.21.1
packaging==23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
django-treebeard~=4.7.1
django-filter~=24.1